refactor: one registry for manifest conformance
Five standalone scripts (~3 400 lines) all asked the same shape of question —
"does this DECLARATION match reality" — for a different declaration each.
They grew one incident at a time and were never designed; nothing connected
"we added a field to the manifest contract" to "therefore something must
verify it", so a new declaration kind could ship with nothing checking it and
you could only find out by reading five files and noticing an absence.

They are now ONE RULE REGISTRY. A rule declares the manifest fields it
governs, and the new `field-coverage` rule DERIVES the set of `gjsify.*` keys
actually declared across the tree and fails on any key no rule claims. Adding a
declaration without a check is no longer forgettable — it is a red build naming
the file to edit. The honest escape is a key -> REASON ledger
(`scripts/manifest-conformance/unchecked-fields.mjs`): the reason is mandatory,
every entry prints on every run, and an entry becomes a failure the moment a
rule claims the key or the field stops being declared — the
`gjsify.platformsUncommitted` idiom, for the same reason.

Running it today surfaces SEVEN uncovered declaration kinds, three of them real
findings (`gjsify.example`, `gjsify.main`, `gjsify.storybook`). Per the brief
none are fixed here; each is deferred with a written reason and tracked in
STATUS.md.

SCOPE is the axis that decides where a rule lives, and it is not a quality
judgement:

  portable (`packages/infra/manifest-conformance/`) reads only the manifest,
  files on disk and binaries — `package-outputs`, `prebuild-artifacts`,
  `headless`, `field-coverage`. Correct in any npm package.

  repo (`scripts/`) knows THIS repository: path-based axis classification,
  curated `@gjsify/*` name allowlists, `prebuilds.yml`'s matrix,
  `resolve-npm`'s alias table, `refs/` submodules — `runtimes-drift`,
  `runtimes-reachability`, `curated-alias-routing`, `tier`, `platforms-ci`,
  `refs-pin`. `runtimes-drift` is the most battle-tested check in the set and
  is firmly repo-scoped: it compares a declaration not to a fact but to a
  heuristic re-derivation built out of one specific tree, so shipping it would
  give a consumer confidently wrong answers.

REGISTRATION IS NOT SELECTION. `package-outputs` and `refs-pin` register (so
coverage sees their fields) but are not selected by `audit-runtimes --check`,
because neither can run in that job — a post-condition on a built tree, and a
check needing initialised submodules.

The CI gate stays a plain Node script. `audit-runtimes.yml` runs with NO
install and NO build; routing it through the CLI would mean booting the
COMMITTED `dist/cli.gjs.mjs`, which reintroduces exactly the staleness
circularity `verify-committed-bundles.mjs` exists to break — a rule added in
source but not rebuilt into the bundle would silently not run, and #821 proved
bundles do merge stale. `verify-committed-bundles.mjs` itself stays wholly
apart: its subject is not a manifest, and it validates the CLI's own artifact.

The shared package has no build step (committed `lib/*.mjs` + a hand-written
`index.d.ts`, the `@gjsify/resolve-npm` shape) — that is what lets a
zero-install CI script and TypeScript import the same modules.

Behaviour is unchanged. All 12 invocation modes of the two rewritten entry
points produce byte-identical stdout, stderr and exit codes, except for the 8
lines the new coverage guard adds. 56 e2e tests across the four affected suites
pass.

`private: true` for now: `gjsify manifest-check` needs a published package, and
publishing a new `@gjsify/*` name needs the manual first-publish +
Trusted-Publisher bootstrap, which does not belong in a refactor. Tracked in
STATUS.md.
