+++ packages/infra/manifest-conformance/lib/rules/package-outputs.mjs
@@ -1,0 +1,239 @@
+/**
+ * Rule `package-outputs` — after a build, every file a package DECLARES must
+ * exist.
+ *
+ * A `package.json` is a contract with two audiences — the bundler/type-checker
+ * inside the repo and every consumer outside it — and both read it through the
+ * same fields: `main`, `module`, `types`/`typings`, `bin`, and every `exports`
+ * condition. A build that leaves one of those paths unwritten produces a package
+ * that resolves to nothing, and NOTHING else in the pipeline notices:
+ *
+ *   - `gjsify tsc` exits 0. With `composite`/`incremental`, a `.tsbuildinfo`
+ *     that outlives its output tree makes tsc consider the project up to date
+ *     and emit NOTHING — a silent no-op with a success exit code. That is issue
+ *     #67, reproduced verbatim in `@gjsify/adwaita-core`: with a stale
+ *     `tmp/.tsbuildinfo` present `gjsify tsc` printed nothing, exited 0, and
+ *     `lib/types/` was never created; deleting the build info made the SAME
+ *     command emit it.
+ *   - The build-cache (ADR 0006) restores per output UNIT. A restore that
+ *     replaces `lib/esm` while `lib/types` is missing is, from the cache's point
+ *     of view, a hit.
+ *   - `gjsify pack`'s type-shipping guard (#655, `tests/e2e/publish-types-shipped`)
+ *     deliberately does NOT fire on an ABSENT declaration — an unbuilt dev tree
+ *     must still pack. It only catches a file that exists but is excluded by
+ *     `files`.
+ *   - `gjsify tsc --noEmit` (the `check` job) type-checks SOURCES. It never
+ *     looks at `lib/`.
+ *
+ * So the only signal was "the build exited 0", and it lied. The consumer-side
+ * symptom is `TS7016: Could not find a declaration file for module
+ * '@gjsify/<pkg>'` — which reads like a consumer bug and gets papered over with
+ * a per-consumer tsconfig workaround.
+ *
+ * It is the same shape as the committed-bundle guard: state the artifact
+ * contract, then check it, rather than trusting an exit code. It is
+ * cause-agnostic — it catches a stale build info, a cache restore that dropped a
+ * unit, a script that was never wired into `build`, and a plain typo in
+ * `exports`, without knowing which one happened.
+ *
+ * WHAT IT COSTS
+ *   - It is a POST-condition, so it needs a built tree: run it AFTER a build,
+ *     never on a fresh clone (`ctx.allowUnbuilt` degrades it to a note for local
+ *     use). In CI that means the `build` job and only there.
+ *   - It proves EXISTENCE, not freshness or correctness. A `.d.ts` regenerated
+ *     from stale sources still passes; that class belongs to the
+ *     committed-bundle guard.
+ *   - Wildcard subpaths (`./assets/*`) can only be checked down to their static
+ *     directory prefix.
+ *   - It reads the declaration as authoritative. A package that declares an
+ *     entry point it genuinely does not ship must fix the declaration — there is
+ *     deliberately no per-package opt-out, because an opt-out list is the thing
+ *     that drifts and then lies.
+ *
+ * PORTABLE: reads nothing but the manifest and the filesystem. The name
+ * exclusions (`@girs/*`, generated type packages, …) are supplied by the CALLER
+ * via `ctx.options.packageOutputs.excludeNamePatterns`, so the rule itself
+ * carries no repository knowledge and behaves identically in a consumer's tree.
+ */
+
+import { existsSync, statSync } from 'node:fs';
+import { posix, relative, resolve } from 'node:path';
+
+import { defineRule } from '../registry.mjs';
+
+/**
+ * A value is a path this rule can check only when it is RELATIVE. Bare
+ * specifiers (`browser` remaps a dependency name), `false` (browser exclusion),
+ * `null` (an `exports` block) and non-strings are somebody else's contract.
+ */
+export function isCheckablePath(value) {
+    if (typeof value !== 'string' || value.length === 0) return false;
+    if (value.startsWith('#') || value.startsWith('node:')) return false;
+    // `./x`, `../x`, `x/y.js`, `x.js` — but not `pkg` or `@scope/pkg`.
+    if (value.startsWith('./') || value.startsWith('../')) return true;
+    return /^[^@][^:]*\.[a-z0-9]+$/i.test(value) || value.includes('/');
+}
+
+/**
+ * Collect `{ field, value }` for every path a package.json declares.
+ * `exports` is walked recursively: condition objects, subpath maps and the
+ * array fallback form all bottom out in strings.
+ */
+export function declaredPaths(pkg) {
+    const out = [];
+
+    for (const field of ['main', 'module', 'types', 'typings', 'style', 'unpkg']) {
+        if (isCheckablePath(pkg[field])) out.push({ field, value: pkg[field] });
+    }
+    // `browser` is either a single entry point or a remap table whose values may
+    // be `false` or another module — only relative values are ours to check.
+    if (typeof pkg.browser === 'string' && isCheckablePath(pkg.browser)) {
+        out.push({ field: 'browser', value: pkg.browser });
+    } else if (pkg.browser && typeof pkg.browser === 'object') {
+        for (const [key, value] of Object.entries(pkg.browser)) {
+            if (isCheckablePath(value)) out.push({ field: `browser[${JSON.stringify(key)}]`, value });
+        }
+    }
+
+    if (typeof pkg.bin === 'string' && isCheckablePath(pkg.bin)) {
+        out.push({ field: 'bin', value: pkg.bin });
+    } else if (pkg.bin && typeof pkg.bin === 'object') {
+        for (const [name, value] of Object.entries(pkg.bin)) {
+            if (isCheckablePath(value)) out.push({ field: `bin[${JSON.stringify(name)}]`, value });
+        }
+    }
+
+    // gjsify's own GJS-first bin map (`gjsify.bin`) is resolved by the CLI the
+    // same way npm resolves `bin`, so it carries the same promise.
+    const gjsifyBin = pkg.gjsify?.bin;
+    if (typeof gjsifyBin === 'string' && isCheckablePath(gjsifyBin)) {
+        out.push({ field: 'gjsify.bin', value: gjsifyBin });
+    } else if (gjsifyBin && typeof gjsifyBin === 'object') {
+        for (const [name, value] of Object.entries(gjsifyBin)) {
+            if (isCheckablePath(value)) out.push({ field: `gjsify.bin[${JSON.stringify(name)}]`, value });
+        }
+    }
+
+    const walkExports = (node, path) => {
+        if (node === null || node === undefined) return;
+        if (typeof node === 'string') {
+            if (isCheckablePath(node)) out.push({ field: path, value: node });
+            return;
+        }
+        if (Array.isArray(node)) {
+            node.forEach((entry, i) => walkExports(entry, `${path}[${i}]`));
+            return;
+        }
+        if (typeof node === 'object') {
+            for (const [key, value] of Object.entries(node)) walkExports(value, `${path}[${JSON.stringify(key)}]`);
+        }
+    };
+    walkExports(pkg.exports, 'exports');
+
+    return out;
+}
+
+/**
+ * What must exist on disk for a declared value.
+ *
+ * A subpath PATTERN (`./assets/*`) has no single target, so the promise it can
+ * still be held to is that the directory it globs into exists — a package that
+ * exports `./assets/*` and ships no `assets/` is broken in exactly the way this
+ * check is for.
+ */
+export function targetFor(pkgDir, value) {
+    if (!value.includes('*')) return { abs: resolve(pkgDir, value), kind: 'file' };
+    const staticPrefix = value.slice(0, value.indexOf('*'));
+    const dir = posix.dirname(`${staticPrefix}x`);
+    return { abs: resolve(pkgDir, dir), kind: 'dir' };
+}
+
+/**
+ * Inspect every in-scope package and report which declared paths are missing.
+ * Exported separately from the rule so the standalone entry script can render
+ * its own `--json` payload from the same data.
+ */
+export function inspectDeclaredOutputs(ctx) {
+    const opts = ctx.options?.packageOutputs ?? {};
+    const excludeNamePatterns = opts.excludeNamePatterns ?? [];
+    const includePrivate = opts.includePrivate === true;
+    const results = [];
+
+    for (const pkg of ctx.packages) {
+        if (ctx.only.length === 0) {
+            if (excludeNamePatterns.some((re) => re.test(pkg.name))) continue;
+            if (pkg.private && !includePrivate) continue;
+        }
+
+        const missing = [];
+        for (const { field, value } of declaredPaths(pkg.manifest)) {
+            const { abs, kind } = targetFor(pkg.dir, value);
+            if (existsSync(abs) && (kind === 'file' || statSync(abs).isDirectory())) continue;
+            missing.push({ field, value, path: relative(ctx.root, abs), kind });
+        }
+        results.push({
+            dir: pkg.rel,
+            name: pkg.name,
+            missing,
+            // Printed on failure so the reader knows what to re-run, without
+            // this rule needing a map of which script owns which output.
+            scripts: Object.fromEntries(
+                Object.entries(pkg.manifest.scripts ?? {}).filter(
+                    ([key]) => key === 'build' || key.startsWith('build:'),
+                ),
+            ),
+        });
+    }
+    return results;
+}
+
+export const packageOutputsRule = defineRule({
+    id: 'package-outputs',
+    scope: 'portable',
+    fields: ['main', 'module', 'types', 'typings', 'style', 'unpkg', 'browser', 'bin', 'exports', 'gjsify.bin'],
+    description: 'every path a package.json declares (main/exports/types/bin/gjsify.bin) exists on disk',
+    run(ctx) {
+        const results = inspectDeclaredOutputs(ctx);
+        const broken = results.filter((r) => r.missing.length > 0);
+        const totalMissing = broken.reduce((n, r) => n + r.missing.length, 0);
+
+        const failures = [];
+        for (const r of broken) {
+            const lines = [`${r.name} (${r.dir}) declares ${r.missing.length} path(s) that do not exist:`];
+            for (const m of r.missing) lines.push(`    ${m.field} → ${m.value}   (missing ${m.kind}: ${m.path})`);
+            const scripts = Object.keys(r.scripts);
+            if (scripts.length > 0) {
+                lines.push(`    produced by: ${scripts.map((s) => `gjsify workspace ${r.name} ${s}`).join(' · ')}`);
+            }
+            failures.push(lines.join('\n'));
+        }
+
+        if (failures.length > 0 && ctx.allowUnbuilt) {
+            return {
+                failures: [],
+                notes: [
+                    `${totalMissing} declared path(s) missing across ${broken.length} package(s) — ` +
+                        'tolerated (--allow-unbuilt). Run a full build and re-check.',
+                ],
+                stats: { checked: results.length, broken: broken.length, totalMissing },
+                results,
+            };
+        }
+
+        return {
+            failures,
+            notes:
+                failures.length > 0
+                    ? [
+                          'A build that leaves a declared entry point unwritten still exits 0 — most often a ' +
+                              '`.tsbuildinfo` that outlived its output tree, which makes `gjsify tsc` a silent no-op. ' +
+                              "Delete the package's build info (its `clear` script does) and rebuild, then fix the " +
+                              'declaration or the build script so it cannot recur.',
+                      ]
+                    : [],
+            summary: `package-outputs: OK. ${results.length} package(s) checked; every declared entry point exists.`,
+            stats: { checked: results.length, broken: broken.length, totalMissing },
+            results,
+        };
+    },
+});
